pyproject.toml
[tool.poetry]
name = "platon-core"
version = "0.0.1"
description = ""
authors = ["Aleksandr Parkhomin <a.parkhomin@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.11.1"

fastapi = "^0.103.2"
greenlet = "2.0.2"

uvicorn = "^0.23.2"
uvloop = "^0.17.0"
databases = "^0.8.0"
psycopg2-binary = "^2.9.8"
asyncpg = "^0.28.0"
pydantic = "^2.4.1"
pyhocon = "^0.3.60"
yoyo-migrations = "^7.0.2"
PyJWT = "^2.8.0"


[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.0"
typing-inspect = '0.6.0'

pytest = "^7.0.0"
pytest-cov = "^4.1.0"
Faker = "^19.6.2"
httpx = "^0.25.0"
mypy = "^1.5.1"
black = "^23.9.1"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
